refactor(database.hosting): fold Database.Server into Database.Hosting [L03.02.01.02.02.01]
The database server runtime is a hosting concern (2026-07-12 owner decision,
mirroring the Web-area Web.Server->Web.Hosting direction): merge the
Database.Server project into Database.Hosting and delete it. Supersedes the
PR #901 DatabaseServer.CreateHostService cross-assembly seam, which existed
only because COHRES002 kept Hosting from referencing Server.

- Server sources move to Database.Hosting (Abstractions/Internal shape kept),
  renamespaced Assimalign.Cohesion.Database.Server -> .Database.Hosting
  (internals join .Hosting.Internal per the module's existing convention).
- COHRES002: the merged csproj declares the sanctioned per-project exemption
  CohesionHostingIsolationExemptions for Assimalign.Cohesion.Database.Protocol
  + Assimalign.Cohesion.Database.Security - the server's own machinery (wire
  framing, authenticator seam), not hosted features. Execution/Types stay
  transitive via the sanctioned area-root reference.
- Composition simplified: DatabaseApplicationOptions.Server carries the
  IDatabaseServer; DatabaseApplication wraps it in the internal endpoint host
  service, registered last (starts last, drains first). CreateHostService is
  internalized; DatabaseServer.Create stays public for manual composition.
- The 16 server tests join Hosting/tests (22 total); Client + Sql.Client test
  csprojs/harnesses now reference Database.Hosting.
- Ripples: App.Database framework manifest (Server entry removed),
  resource-database.yml matrix 19->18, three slnx files, area README/DESIGN,
  Hosting docs absorb the Server design record (session state machine,
  two-phase drain, error mapping, bound-listener decision), and cross-project
  doc references updated.

Verified: Hosting 22, Client 11, Sql.Client 12 tests green; App.Database
Runtime + Ref packs green (Release).
