+++ docs/DATABASE_PROGRAM_PLAN.md
@@ -65,7 +65,7 @@ A stage is a gate, not a calendar. Everything in a stage may proceed once the pr
 | **K — Storage kernel** | pages, WAL, recovery, backup | `Database.Storage` | The journal is the *only* durability mechanism; no side files. Page CRC on every read path. Keep contracts model-agnostic — model layouts live in `{Model}.Storage`, never here. |
 | **T — ACID substrate** | transactions, locks, versions, types, indexing | `Database.Transactions`, `Database.Types`, `Database.Indexing` | `TransactionSequence` (ordering) vs `TransactionId` (identity) — don't conflate. Index mutations always take `ITransactionContext`. Type encodings must be order-preserving under raw byte compare. |
 | **X — Execution** | sessions, pipeline, plans | `Database.Execution`, root `Database` | Model-agnostic: no SQL/document semantics in shared operators. `QueryRequest`/`QueryResult` families stay abstract; models subclass. |
-| **P — Protocol & clients** | wire protocol, server, client core | `Database.Protocol`, `Database.Server`, `Database.Client` | Framing is model-agnostic; model semantics live in payloads. Every limit (payload size, sessions, timeouts) is DoS-critical — enforce at parse/accept. Test over `Connections.InMemory`, never live sockets. |
+| **P — Protocol & clients** | wire protocol, server runtime, client core | `Database.Protocol`, `Database.Hosting` (server runtime; `Database.Server` folded in 2026-07-12), `Database.Client` | Framing is model-agnostic; model semantics live in payloads. Every limit (payload size, sessions, timeouts) is DoS-critical — enforce at parse/accept. Test over `Connections.InMemory`, never live sockets. |
 | **LS / LD / LG — Languages** | per-model parsers | `Sql.Language`, `Documents.Language`, `Graph.Language` | Always parallel-safe (leaf dependencies). Declared-dialect contract before deep implementation; conformance corpora guard the AST. Graph language work gates on the #193 standard decision. |
 | **MS / MD / MK / MB / MG — Model engines** | engine + model storage + catalog | `Database.{Sql,Documents,KeyValuePair,Blob,Graph}` + their `.Storage`/`.Catalog` | Compose the kernel — never re-implement paging/WAL/locking locally. Every write path goes through `ITransactionContext`. Engines never depend on other engines. |
 | **H — Hosting & ops** | host, governance, health, orchestration | `Database.Hosting`, `Database.Governance`, `Database.ApplicationModel` | WAL flush / page writer = `DedicatedThreadService`; endpoint = `BackgroundService` (Hosting DESIGN.md menu). `*.Hosting` is the ONLY DI/Config/Logging seam. The manifest project never references the runtime. |
