+++ resources/Assimalign.Cohesion.Resources.slnx
@@ -391,17 +391,6 @@
 	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Security/src/">
 		<Project Path="Database/Assimalign.Cohesion.Database.Security/src/Assimalign.Cohesion.Database.Security.csproj" />
 	</Folder>
-	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Server/" />
-	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Server/docs/">
-		<File Path="Database/Assimalign.Cohesion.Database.Server/docs/DESIGN.md" />
-		<File Path="Database/Assimalign.Cohesion.Database.Server/docs/OVERVIEW.md" />
-	</Folder>
-	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Server/src/">
-		<Project Path="Database/Assimalign.Cohesion.Database.Server/src/Assimalign.Cohesion.Database.Server.csproj" />
-	</Folder>
-	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Server/tests/">
-		<Project Path="Database/Assimalign.Cohesion.Database.Server/tests/Assimalign.Cohesion.Database.Server.Tests.csproj" />
-	</Folder>
 	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Sql/" />
 	<Folder Name="/resources/Database/Assimalign.Cohesion.Database.Sql/src/">
 		<Project Path="Database/Assimalign.Cohesion.Database.Sql/src/Assimalign.Cohesion.Database.Sql.csproj" />
