+++ resources/Database/Assimalign.Cohesion.Database.Client/tests/Assimalign.Cohesion.Database.Client.Tests.csproj
@@ -9,9 +9,10 @@
 		<CohesionPackageReference Include="xunit.runner.visualstudio" />
 		<CohesionPackageReference Include="Shouldly" />
 		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Client" />
-		<!-- E2E acceptance: the real server + a live SQL engine over the in-memory
-		     Connections driver (tests are exempt from the area layering rules). -->
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Server" />
+		<!-- E2E acceptance: the real server (owned by Database.Hosting) + a live SQL
+		     engine over the in-memory Connections driver (tests are exempt from the
+		     area layering rules). -->
+		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Hosting" />
 		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Sql" />
 		<CohesionProjectReference Include="Assimalign.Cohesion.Connections.InMemory" />
 	</ItemGroup>
