+++ resources/Database/Assimalign.Cohesion.Database.Hosting/src/DatabaseServer.cs
@@ -1,8 +1,8 @@
 using System;
 
-using Assimalign.Cohesion.Hosting;
+using Assimalign.Cohesion.Database.Hosting.Internal;
 
-namespace Assimalign.Cohesion.Database.Server;
+namespace Assimalign.Cohesion.Database.Hosting;
 
 /// <summary>
 /// Creates database server instances from options.
@@ -17,6 +17,13 @@ public static class DatabaseServer
     /// <returns>The server.</returns>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="options"/> is null.</exception>
     /// <exception cref="ArgumentException">Thrown when the options carry no listener, no engines, or a non-positive session limit.</exception>
+    /// <remarks>
+    /// <see cref="DatabaseApplication"/> composes and runs the created server as its
+    /// endpoint host service (<see cref="DatabaseApplicationOptions.Server"/>). A
+    /// custom or embedded host composes the server manually and drives
+    /// <see cref="IDatabaseServer.StartAsync"/>/<see cref="IDatabaseServer.StopAsync"/>
+    /// on its own lifecycle.
+    /// </remarks>
     public static IDatabaseServer Create(DatabaseServerOptions options)
     {
         ArgumentNullException.ThrowIfNull(options);
@@ -35,25 +42,5 @@ public static class DatabaseServer
         }
 
         return new DefaultDatabaseServer(options);
-    }
-
-    /// <summary>
-    /// Wraps a server as an <see cref="IHostService"/> so a host can run its accept
-    /// loop and drain it on shutdown per the hosting execution menu.
-    /// </summary>
-    /// <param name="server">The server to host.</param>
-    /// <returns>An endpoint host service that starts the server on start and gracefully drains it on stop.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="server"/> is null.</exception>
-    /// <remarks>
-    /// The adapter lives here rather than in <c>Database.Hosting</c> because the
-    /// resource hosting-isolation rule (COHRES002) bars the hosting module from
-    /// referencing any same-area library except the area root — so only the server's
-    /// own package can name <see cref="IDatabaseServer"/> to compose it. The hosting
-    /// module composes the returned <see cref="IHostService"/> generically.
-    /// </remarks>
-    public static IHostService CreateHostService(IDatabaseServer server)
-    {
-        ArgumentNullException.ThrowIfNull(server);
-        return new DatabaseServerHostService(server);
     }
 }
