+++ resources/Database/Assimalign.Cohesion.Database.Hosting/tests/Assimalign.Cohesion.Database.Hosting.Tests.csproj
@@ -9,10 +9,9 @@
 		<CohesionPackageReference Include="xunit.runner.visualstudio" />
 		<CohesionPackageReference Include="Shouldly" />
 		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Hosting" />
-		<!-- Composition acceptance: the host runs a real SQL engine + wire server over
-		     the in-memory driver, driven end-to-end by a client (tests are exempt from
-		     the area layering rules). -->
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Server" />
+		<!-- Composition acceptance: the host runs a real SQL engine + the wire server
+		     (owned by Hosting) over the in-memory driver, driven end-to-end by a
+		     client (tests are exempt from the area layering rules). -->
 		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Sql" />
 		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Client" />
 		<CohesionProjectReference Include="Assimalign.Cohesion.Connections.InMemory" />
