+++ resources/Database/Assimalign.Cohesion.Database.Hosting/tests/TestObjects/DatabaseHostTestHarness.cs
@@ -5,7 +5,6 @@ using System.Threading.Tasks;
 using Assimalign.Cohesion.Connections.InMemory;
 using Assimalign.Cohesion.Database.Client;
 using Assimalign.Cohesion.Database.Hosting;
-using Assimalign.Cohesion.Database.Server;
 using Assimalign.Cohesion.Database.Sql;
 using Assimalign.Cohesion.Hosting;
 
@@ -59,7 +58,7 @@ internal sealed class DatabaseHostTestHarness : IAsyncDisposable
 
         var options = new DatabaseApplicationOptions();
         options.Engines.Add(engine);
-        options.Services.Add(DatabaseServer.CreateHostService(server));
+        options.Server = server;
         configure?.Invoke(options);
 
         var application = new DatabaseApplication(options);
