+++ resources/Database/Assimalign.Cohesion.Database.Hosting/tests/TestObjects/ProtocolTestClient.cs
@@ -6,7 +6,7 @@ using Shouldly;
 using Assimalign.Cohesion.Connections;
 using Assimalign.Cohesion.Database.Protocol;
 
-namespace Assimalign.Cohesion.Database.Server.Tests;
+namespace Assimalign.Cohesion.Database.Hosting.Tests;
 
 /// <summary>
 /// A hand-driven protocol client over one in-memory connection: sends raw frames
