+++ resources/Database/Assimalign.Cohesion.Database.Hosting/tests/TestObjects/RejectingAuthenticator.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 using Assimalign.Cohesion.Database.Security;
 
-namespace Assimalign.Cohesion.Database.Server.Tests;
+namespace Assimalign.Cohesion.Database.Hosting.Tests;
 
 /// <summary>
 /// An authenticator that rejects every principal, for handshake-failure tests.
