+++ resources/Database/Assimalign.Cohesion.Database.Hosting/tests/TestObjects/TestTimeout.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading;
 
-namespace Assimalign.Cohesion.Database.Server.Tests;
+namespace Assimalign.Cohesion.Database.Hosting.Tests;
 
 /// <summary>
 /// Bounded cancellation tokens so a hung exchange fails the test instead of the
