--- resources/Database/Assimalign.Cohesion.Database.Server/src/Assimalign.Cohesion.Database.Server.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-	<ItemGroup>
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Execution" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Protocol" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Security" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Types" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Connections" />
-		<!-- Non-area hosting foundation (libraries/Hosting, NOT the area's *.Hosting
-		     module): supplies the BackgroundService base for the endpoint host-service
-		     adapter that Database.Hosting composes. See docs/DESIGN.md (Placement). -->
-		<CohesionProjectReference Include="Assimalign.Cohesion.Hosting" />
-	</ItemGroup>
-</Project>
