--- resources/Database/Assimalign.Cohesion.Database.Server/tests/Assimalign.Cohesion.Database.Server.Tests.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<IsPackable>false</IsPackable>
-	</PropertyGroup>
-	<ItemGroup>
-		<CohesionPackageReference Include="Microsoft.NET.Test.Sdk" />
-		<CohesionPackageReference Include="coverlet.collector" />
-		<CohesionPackageReference Include="xunit" />
-		<CohesionPackageReference Include="xunit.runner.visualstudio" />
-		<CohesionPackageReference Include="Shouldly" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Server" />
-		<!-- E2E acceptance: a live SQL engine over the in-memory Connections driver
-		     (tests are exempt from the area layering rules). -->
-		<CohesionProjectReference Include="Assimalign.Cohesion.Database.Sql" />
-		<CohesionProjectReference Include="Assimalign.Cohesion.Connections.InMemory" />
-	</ItemGroup>
-</Project>
