+++ resources/Database/Assimalign.Cohesion.Database.Sql.Client/tests/TestObjects/SqlClientTestHarness.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 using Assimalign.Cohesion.Connections.InMemory;
 using Assimalign.Cohesion.Database.Client;
-using Assimalign.Cohesion.Database.Server;
+using Assimalign.Cohesion.Database.Hosting;
 using Assimalign.Cohesion.Database.Sql;
 
 namespace Assimalign.Cohesion.Database.Sql.Client.Tests;
