+++ resources/Database/Assimalign.Cohesion.Database.slnx
@@ -350,17 +350,6 @@
 	<Folder Name="/Database/Assimalign.Cohesion.Database.Security/src/">
 		<Project Path="Assimalign.Cohesion.Database.Security/src/Assimalign.Cohesion.Database.Security.csproj" />
 	</Folder>
-	<Folder Name="/Database/Assimalign.Cohesion.Database.Server/" />
-	<Folder Name="/Database/Assimalign.Cohesion.Database.Server/docs/">
-		<File Path="Assimalign.Cohesion.Database.Server/docs/DESIGN.md" />
-		<File Path="Assimalign.Cohesion.Database.Server/docs/OVERVIEW.md" />
-	</Folder>
-	<Folder Name="/Database/Assimalign.Cohesion.Database.Server/src/">
-		<Project Path="Assimalign.Cohesion.Database.Server/src/Assimalign.Cohesion.Database.Server.csproj" />
-	</Folder>
-	<Folder Name="/Database/Assimalign.Cohesion.Database.Server/tests/">
-		<Project Path="Assimalign.Cohesion.Database.Server/tests/Assimalign.Cohesion.Database.Server.Tests.csproj" />
-	</Folder>
 	<Folder Name="/Database/Assimalign.Cohesion.Database.Sql/" />
 	<Folder Name="/Database/Assimalign.Cohesion.Database.Sql/src/">
 		<Project Path="Assimalign.Cohesion.Database.Sql/src/Assimalign.Cohesion.Database.Sql.csproj" />
