+++ resources/Database/README.md
@@ -39,12 +39,11 @@ Each model follows the same matrix: root (engine + public interface), plus `.Lan
 | Project | Role |
 |---|---|
 | `Assimalign.Cohesion.Database.Protocol` | Wire protocol frames and message contracts (shared client/server) |
-| `Assimalign.Cohesion.Database.Server` | Network front-end: sessions, auth handshake, frame pump |
 | `Assimalign.Cohesion.Database.Client` | Shared client core: connection strings, pooling, protocol client |
 | `Assimalign.Cohesion.Database.Security` | AuthN/AuthZ contracts (principals, roles, permissions) |
 | `Assimalign.Cohesion.Database.Replication` | Shared replication contracts (WAL log-shipping seam) |
 | `Assimalign.Cohesion.Database.Governance` | Quotas, tenancy boundaries, audit events |
-| `Assimalign.Cohesion.Database.Hosting` | Host composition (`Host<TContext>`), the area's only DI seam |
+| `Assimalign.Cohesion.Database.Hosting` | Host composition (`Host<TContext>`), the area's only DI seam; owns the server runtime — network front-end, sessions, auth handshake, frame pump (`Database.Server` folded in 2026-07-12) |
 | `Assimalign.Cohesion.Database.ApplicationModel` | Manifest-only orchestration resource + `AddDatabase(...)` |
 | `Assimalign.Cohesion.Database.Embedded` | In-process consumption facade — how other platform resources embed their data layer |
 
@@ -52,7 +51,7 @@ Each model follows the same matrix: root (engine + public interface), plus `.Lan
 
 - `libraries/Core` — foundational primitives (everywhere)
 - `libraries/Hosting` — host lifecycle + per-service execution menu (`Database.Hosting`)
-- `libraries/Connections` — transport drivers for the server (`Database.Server`)
+- `libraries/Connections` — transport drivers for the server runtime (`Database.Hosting`)
 - `libraries/ApplicationModel` — orchestration contracts (`Database.ApplicationModel` only)
 - `resources/Web` — private implementation detail of the root project (HTTP admin surface); hidden from consumers via `CohesionPrivateProjectReference`
